refactor: domain, util 패키지 분리
- validator를 생성하면서 util 패키지를 정의했는데, 계획된 흐름을 벗어난 것 같아 개선이 필요하다.
- 패키지 구조를 재구성하면서 도메인 객체를 별도로 묶었다. Cars 같은 경우는 wrapper에 속하나, 도메인과 래퍼 클래스의 경계가 모호해져서 유지했다.
